+++ packages/core_event_source_test_util/lib/core_event_source_test_util.dart
@@ -2,4 +2,10 @@
 
 library core_event_source_test_util;
 
+export 'src/bloc_test2.dart';
 export 'src/data_store_specification.dart';
+export 'src/fake_behavior.dart';
+export 'src/integration_test_help.dart';
+export 'src/t.dart';
+export 'src/test_group/execute_commands_test.dart';
+export 'src/test_group/forward_and_follow_test_group.dart';
