+++ packages/core_event_source_test_util/lib/src/fake_behavior.dart
