+++ packages/core_event_source_test_util/lib/src/test_group/forward_and_follow_test_group.dart
@@ -1,0 +1,49 @@
+import 'dart:math';
+
+import 'package:bloc_test/bloc_test.dart';
+import 'package:core_event_source/event_source.dart';
+import 'package:test/test.dart';
+
+import '../../core_event_source_test_util.dart';
+
+class ForwardAndFollowTestGroup extends IntegrationTestGroup {
+  ForwardAndFollowTestGroup(super.eventStoreFactory);
+
+  static const headRefName1 = '1';
+  static const headRefName2 = '2';
+
+  @override
+  run() {
+    late EventStore eventStore;
+    late String sourcePath;
+    late EventSource<FakeCommand, FakeView> source1;
+    late EventSource<FakeCommand, FakeView> source2;
+
+    group('Execute Commands', () {
+      setUp(() async => eventStore = await eventStoreFactory.create());
+      setUp(() => sourcePath = 'objects/${Random.secure().nextDouble()}');
+      setUp(() async =>
+          source1 = await build(eventStore, sourcePath, headRefName1));
+
+      blocTest2(
+        'execute commands on separate headRef instances, in sequence',
+        setUp: () async {
+          await source1.isReady;
+          final done = source1.stream.take(1).first;
+          source1.execute([{}]);
+          await done;
+          await source1.close();
+          source2 = await build(eventStore, sourcePath, headRefName2);
+        },
+        build: () => source2,
+        act: (_) async {
+          await source2.isReady;
+          source2.execute([{}]);
+        },
+        wait: const Duration(milliseconds: 10),
+        expect: () => [1, 2, 4],
+        tearDown: () async => await source2.close(),
+      );
+    });
+  }
+}
