+++ packages/flutter_core_event_source/integration_test/integration_test.dart
@@ -1,16 +1,17 @@
+import 'package:flutter_core_event_source_test_util/flutter_core_event_source_test_util.dart';
 import 'package:integration_test/integration_test.dart';
 
 import '../test/functional/execute_commands_test.dart' as execute_commands_test;
-import '../test/util/logging.dart';
+import '../test/functional/forward_and_follow_test.dart'
+    as forward_and_follow_test;
 
 main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
 
   initializeDebugLogging();
   final mains = [
-    // firestore_event_store_test.main,
     execute_commands_test.main,
-    // event_source_impl_test.main
+    forward_and_follow_test.main,
   ];
   for (final main in mains) {
     main();
