--- packages/flutter_core_event_source/test/test_doubles.dart
@@ -1,5 +1,0 @@
-// GENERATED CODE - DO NOT MODIFY BY HAND
-
-library test_doubles;
-
-export 'test_doubles/fake_behavior.dart';
