--- packages/flutter_core_event_source/test/test_util.dart
@@ -1,10 +1,0 @@
-// GENERATED CODE - DO NOT MODIFY BY HAND
-
-library test_util;
-
-export 'util/debug_bloc_observer.dart';
-export 'util/event_sourced_behavior_fake.dart';
-export 'util/fake_firebase_firestore.dart';
-export '../integration_test/firebase_options.dart';
-export 'util/logging.dart';
-export 'util/t.dart';
