--- packages/flutter_core_event_source/test/util/event_sourced_behavior_fake.dart
@@ -1,15 +1,0 @@
-import 'package:core_event_source/event_sourced_behavior.dart';
-import 'package:flutter_test/flutter_test.dart';
-
-class EventSourcedBehaviorFake<Command, Event, State, View> extends Fake
-    implements EventSourcedBehavior<Command, Event, State, View> {
-  @override
-  final CommandHandler<Command, Event, State> commandHandler;
-  @override
-  final StateEventHandler<Event, State> eventHandler;
-
-  EventSourcedBehaviorFake({
-    required this.commandHandler,
-    required this.eventHandler,
-  });
-}
