+++ packages/flutter_core_event_source_test_util/.github/workflows/main.yaml
@@ -1,0 +1,26 @@
+name: ci
+
+concurrency:
+  group: ${{ github.workflow }}-${{ github.ref }}
+  cancel-in-progress: true
+
+on:
+  pull_request:
+    branches:
+      - main
+
+jobs:
+  semantic_pull_request:
+    uses: ./.github/workflows/semantic_pull_request.yml
+
+  spell-check:
+    uses: ./.github/workflows/spell_check.yml
+    with:
+      includes: "**/*.md"
+      modified_files_only: false
+
+  build:
+    uses: ./.github/workflows/flutter_package.yml
+    with:
+      flutter_channel: stable
+      min_coverage: 90
