+++ packages/flutter_core_event_source_test_util/CHANGELOG.md
@@ -1,0 +1,3 @@
+# 0.1.0+1
+
+- feat: initial commit 🎉
