+++ packages/flutter_core_event_source_test_util/analysis_options.yaml
@@ -1,0 +1,4 @@
+include: package:lints/core.yaml
+analyzer:
+  exclude:
+    - lib/src/entry/entry.g.dart
