+++ packages/flutter_core_event_source_test_util/lib/flutter_core_event_source_test_util.dart
@@ -1,0 +1,8 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+library flutter_core_event_source_test_util;
+
+export 'src/debug_bloc_observer.dart';
+export 'src/fake_firebase_firestore.dart';
+export 'src/firestore_test_event_store_factory.dart';
+export 'src/logging.dart';
