+++ packages/flutter_core_event_source_test_util/lib/src/logging.dart
