+++ packages/flutter_core_event_source_test_util/pubspec.yaml
@@ -1,21 +1,27 @@
-name: core_event_source_test_util
+name: flutter_core_event_source_test_util
 description: A very good flutter package
 version: 0.1.0+1
 publish_to: none
 
 environment:
   sdk: ">=2.19.0 <3.0.0"
+  flutter: 3.7.0
 
 dependencies:
+  flutter:
+    sdk: flutter
   core_event_source:
     path: ../core_event_source
-#    git:
-#      url: https://github.com/ryanhanks-bestow/core_event_source
-#      ref: main
+  #    git:
+  #      url: https://github.com/ryanhanks-bestow/core_event_source
+  #      ref: main
+  flutter_core_event_source:
+    path: ../flutter_core_event_source
+  core_event_source_test_util:
+    path: ../core_event_source_test_util
 
   async:
   bloc:
-  bloc_test:
   bloc_concurrency:
   collection:
   debounce_throttle:
@@ -28,8 +34,16 @@ dependencies:
   matcher:
   given_when_then_unit_test:
   test:
+  firebase_auth:
+  firebase_core:
+  cloud_firestore:
 
 dev_dependencies:
+  flutter_test:
+    sdk: flutter
+  integration_test:
+    sdk: flutter
+  bloc_test:
   build_runner:
   lints: ^2.0.1
   freezed:
@@ -43,6 +57,6 @@ index_generator:
     - '**.*.dart'
   indexes:
     - path: lib
-      name: core_event_source_test_util
+      name: flutter_core_event_source_test_util
       include:
         - src
